[PAPP-742] Updated all the tests
